+++ lib/nuvigator.dart
@@ -3,6 +3,7 @@ library nuvigator;
 export 'src/annotations.dart';
 export 'src/nuvigator.dart';
 export 'src/router.dart';
+export 'src/router_helpers.dart';
 export 'src/screen_route.dart';
 export 'src/screen_type.dart';
 export 'src/screen_types/cupertino_screen_type.dart';
