+++ test/router_test.dart
@@ -27,7 +27,7 @@ void main() {
 
   test('router retrieves the right screen for the deepLink', () async {
     final testRouter = TestRouter();
-    final routeEntry = testRouter.getRouteEntryForDeepLink('test/simple');
+//    final routeEntry = testRouter.getRouteEntryForDeepLink('test/simple');
     expect(routeEntry.screenBuilder(const RouteSettings()).debugKey,
         'testRouterFirstScreen');
   });
